git-svn-id: http://192.168.0.5/svn/Member/trade/Spring@6586 8c2d0303-d7d2-0546-a2d7-8b6cbec2f4b4

+++ Util/Tool/SolrSearch.php
@@ -129,11 +129,17 @@ class SolrSearch
 	{
 		$this->connect($collection);
 		$this->parse($rule);
+		$this->fq && $this->solrQuery->setQuery(implode(' AND ', $this->fq));
 		$query = $this->solrQuery->getQuery();
-		$this->solrClient->deleteByQuery($query);
-		$this->solrClient->commit();
+		if ( $query ) 
+		{
+			$this->solrClient->deleteByQuery($query);
+			$this->solrClient->commit();
 
-		return true;
+			return true;
+		}
+
+		return false;
 	}
 
 	/**
